Split the mid-air double jump onto its own rebindable button
The ground jump and the mid-air double jump shared the jump key. The double
jump now has a dedicated input (BTN_AIRJUMP); the button mask widens to 16 bits
(protocol v5) to carry it. Wall jumps stay on the primary jump button.

The bind defaults to "same as jump" so out-of-box feel is unchanged, and can be
moved onto its own key/wheel in the settings menu, via --doublejump, or the
saved config (doublejump = jump|space|mwheelup|mwheeldown). Movement replicates
air_jump_buffer/air_jump_held in snapshots so client prediction stays exact.

Includes a client-path regression test: BTN_AIRJUMP (256) must survive the
client's pending-button accumulation, which a leftover uint8_t field truncated.

+++ src/client/client.h
@@ -32,7 +32,7 @@ struct Client {
   double last_recv_time, connect_start_time, last_hello_time;
   double net_now;
   double next_input_send_time;
-  uint8_t pending_buttons;
+  uint16_t pending_buttons;
   uint8_t pending_weapon_switch;
   uint8_t pending_class_switch;
   PlayerInput sent_inputs[CLIENT_INPUT_RING];
